Autenticacion OAuth2 con JWT

+++ spring-boot-backend-apirest/src/main/java/com/capename/springboot/backend/apirest/controllers/ClienteRestController.java
@@ -20,6 +20,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -33,6 +34,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import com.capename.springboot.backend.apirest.models.entity.Cliente;
+import com.capename.springboot.backend.apirest.models.entity.Region;
 import com.capename.springboot.backend.apirest.models.services.IClienteService;
 import com.capename.springboot.backend.apirest.models.services.IUploadFileService;
 
@@ -63,6 +65,7 @@ public class ClienteRestController {
 		
 	}
 	
+	@Secured({ "ROLE_ADMIN", "ROLE_USER" })
 	@GetMapping("/clientes/{id}")	
 	public ResponseEntity<?> show( @PathVariable Long id ) {
 		
@@ -88,6 +91,7 @@ public class ClienteRestController {
 		
 	}
 	
+	@Secured("ROLE_ADMIN")
 	@PostMapping("/clientes")	
 	public ResponseEntity<?> create( @Valid @RequestBody Cliente cliente, BindingResult result) {
 		Cliente clienteNuevo = null; 
@@ -127,6 +131,7 @@ public class ClienteRestController {
 		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED );
 	}
 	
+	@Secured( "ROLE_ADMIN" )
 	@PutMapping("/clientes/{id}")	
 	public ResponseEntity<?> update( @Valid @RequestBody Cliente cliente, BindingResult result ,@PathVariable Long id ) {
 		
@@ -157,6 +162,7 @@ public class ClienteRestController {
 			clienteActual.setNombre(cliente.getNombre());
 			clienteActual.setEmail(cliente.getEmail());	
 			clienteActual.setCreateAt(cliente.getCreateAt());
+			clienteActual.setRegion(cliente.getRegion());
 			
 			clienteActualizado = clienteService.save(clienteActual);
 			
@@ -173,14 +179,20 @@ public class ClienteRestController {
 		return new ResponseEntity<Map<String, Object>>( response, HttpStatus.OK ) ;
 	}
 	
+	@Secured( "ROLE_ADMIN")
 	@DeleteMapping("/clientes/{id}")	
 	public ResponseEntity<?> delete( @PathVariable Long id ) {
 		Map<String, Object> response = new HashMap<>();		
 		
 		try {
-			Cliente cliente = clienteService.findById(id);
+			Cliente cliente = clienteService.findById(id);			
+			if (cliente == null) {
+				response.put("mensaje", "Cliente id: " + id + " no existe!");				
+				return new ResponseEntity<Map<String, Object>>( response, HttpStatus.NOT_FOUND );				
+			}
 			String nombreFotoAnterior = cliente.getFoto();	
 			
+			@SuppressWarnings("unused")
 			boolean isFotoDeleted = uploadService.eliminarFoto(nombreFotoAnterior);			
 			
 			clienteService.delete(id);
@@ -197,11 +209,17 @@ public class ClienteRestController {
 		return new ResponseEntity<Map<String, Object>>( response, HttpStatus.OK );
 	}
 	
+	@Secured({ "ROLE_ADMIN", "ROLE_USER" })
 	@PostMapping("/clientes/upload")
 	public ResponseEntity<?> upload( @RequestParam("archivo") MultipartFile archivo, @RequestParam("id") Long id ){		
 		Map<String, Object> response = new HashMap<>();
 		
-		Cliente cliente = clienteService.findById(id);		
+		Cliente cliente = clienteService.findById(id);	
+		
+		if (cliente == null) {
+			response.put("error", "Cliente id: " + id + " no existe!");				
+			return new ResponseEntity<Map<String, Object>>( response, HttpStatus.NOT_FOUND );				
+		}		
 		
 		if (!archivo.isEmpty()) {	
 			String nombreArchivo = "";
@@ -252,6 +270,13 @@ public class ClienteRestController {
 		return new ResponseEntity<Resource>(recurso, cabeceraHttp, HttpStatus.OK);
 		
 	}
+	
+	@Secured( "ROLE_ADMIN" )
+	@GetMapping("/clientes/regiones")
+	public List<Region> listarRegiones(){
+		return clienteService.findAllRegiones();
+	}
+
